Moved StdTraits into its own .hpp file.
