+++ src/pyco_tree/pico_tree/_pyco_tree/pyco_adaptor.hpp
@@ -2,7 +2,7 @@
 
 #include <pybind11/numpy.h>
 
-#include <pico_tree/core.hpp>
+#include <pico_tree/std_traits.hpp>
 
 #include "core.hpp"
